# backend/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type ActivityLog struct {
	ID         pgtype.UUID        `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	Action     string             `json:"action"`
	TargetType pgtype.Text        `json:"target_type"`
	TargetID   pgtype.UUID        `json:"target_id"`
	Details    []byte             `json:"details"`
	IpAddress  pgtype.Text        `json:"ip_address"`
	UserAgent  pgtype.Text        `json:"user_agent"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type AnggaranDokumenBukti struct {
	ID             pgtype.UUID        `json:"id"`
	AnggaranNodeID pgtype.UUID        `json:"anggaran_node_id"`
	FileHashSha256 string             `json:"file_hash_sha256"`
	OriginalName   string             `json:"original_name"`
	MimeType       string             `json:"mime_type"`
	FileSizeBytes  int64              `json:"file_size_bytes"`
	UploadedBy     pgtype.UUID        `json:"uploaded_by"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	DeletedAt      pgtype.Timestamptz `json:"deleted_at"`
}

type AnggaranHistory struct {
	ID                   pgtype.UUID        `json:"id"`
	AnggaranNodeID       pgtype.UUID        `json:"anggaran_node_id"`
	ParentID             pgtype.UUID        `json:"parent_id"`
	Jenis                string             `json:"jenis"`
	Kode                 string             `json:"kode"`
	Uraian               string             `json:"uraian"`
	TahunAnggaran        pgtype.Int4        `json:"tahun_anggaran"`
	PaguRevisi           pgtype.Numeric     `json:"pagu_revisi"`
	LockPagu             pgtype.Numeric     `json:"lock_pagu"`
	RealisasiPeriodeLalu pgtype.Numeric     `json:"realisasi_periode_lalu"`
	RealisasiPeriodeIni  pgtype.Numeric     `json:"realisasi_periode_ini"`
	RealisasiSdPeriode   pgtype.Numeric     `json:"realisasi_sd_periode"`
	PersentaseRealisasi  pgtype.Numeric     `json:"persentase_realisasi"`
	SisaAnggaran         pgtype.Numeric     `json:"sisa_anggaran"`
	SnapshotPeriode      string             `json:"snapshot_periode"`
	SnapshotCreatedAt    pgtype.Timestamptz `json:"snapshot_created_at"`
	Source               pgtype.Text        `json:"source"`
}

type AnggaranNode struct {
	ID                   pgtype.UUID    `json:"id"`
	ParentID             pgtype.UUID    `json:"parent_id"`
	Jenis                string         `json:"jenis"`
	Kode                 string         `json:"kode"`
	Uraian               string         `json:"uraian"`
	TahunAnggaran        pgtype.Int4    `json:"tahun_anggaran"`
	PaguRevisi           pgtype.Numeric `json:"pagu_revisi"`
	LockPagu             pgtype.Numeric `json:"lock_pagu"`
	RealisasiPeriodeLalu pgtype.Numeric `json:"realisasi_periode_lalu"`
	RealisasiPeriodeIni  pgtype.Numeric `json:"realisasi_periode_ini"`
	RealisasiSdPeriode   pgtype.Numeric `json:"realisasi_sd_periode"`
	PersentaseRealisasi  pgtype.Numeric `json:"persentase_realisasi"`
	SisaAnggaran         pgtype.Numeric `json:"sisa_anggaran"`
	Source               pgtype.Text    `json:"source"`
}

type DipaDokuman struct {
	ID             pgtype.UUID        `json:"id"`
	TahunAnggaran  int32              `json:"tahun_anggaran"`
	Bulan          int32              `json:"bulan"`
	Revisi         int32              `json:"revisi"`
	FileHashSha256 string             `json:"file_hash_sha256"`
	OriginalName   string             `json:"original_name"`
	MimeType       string             `json:"mime_type"`
	FileSizeBytes  int64              `json:"file_size_bytes"`
	UploadedBy     pgtype.UUID        `json:"uploaded_by"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
}

type DokumenBukti struct {
	ID                 pgtype.UUID        `json:"id"`
	PaketID            pgtype.UUID        `json:"paket_id"`
	Bulan              int32              `json:"bulan"`
	Kategori           string             `json:"kategori"`
	JenisDokumen       string             `json:"jenis_dokumen"`
	FileHashSha256     string             `json:"file_hash_sha256"`
	OriginalName       string             `json:"original_name"`
	MimeType           string             `json:"mime_type"`
	FileSizeBytes      int64              `json:"file_size_bytes"`
	UploadedBy         pgtype.UUID        `json:"uploaded_by"`
	CreatedAt          pgtype.Timestamptz `json:"created_at"`
	VerificationStatus pgtype.Text        `json:"verification_status"`
	VerifiedBy         pgtype.UUID        `json:"verified_by"`
	VerifiedAt         pgtype.Timestamptz `json:"verified_at"`
	RejectionReason    pgtype.Text        `json:"rejection_reason"`
}

type PaketAkunMapping struct {
	PaketID pgtype.UUID `json:"paket_id"`
	AkunID  pgtype.UUID `json:"akun_id"`
}

type PaketPekerjaan struct {
	ID        pgtype.UUID        `json:"id"`
	NamaPaket string             `json:"nama_paket"`
	Kasatker  string             `json:"kasatker"`
	Lokasi    string             `json:"lokasi"`
	PaguPaket pgtype.Numeric     `json:"pagu_paket"`
	Status    string             `json:"status"`
	PpkID     pgtype.UUID        `json:"ppk_id"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type PaketRealisasiFisik struct {
	ID                 pgtype.UUID        `json:"id"`
	PaketID            pgtype.UUID        `json:"paket_id"`
	Bulan              int32              `json:"bulan"`
	PersenAktual       pgtype.Numeric     `json:"persen_aktual"`
	CatatanKendala     pgtype.Text        `json:"catatan_kendala"`
	UpdatedBy          pgtype.UUID        `json:"updated_by"`
	CreatedAt          pgtype.Timestamptz `json:"created_at"`
	VerificationStatus pgtype.Text        `json:"verification_status"`
	VerifiedBy         pgtype.UUID        `json:"verified_by"`
	VerifiedAt         pgtype.Timestamptz `json:"verified_at"`
	RejectionReason    pgtype.Text        `json:"rejection_reason"`
}

type PaketTarget struct {
	ID             pgtype.UUID    `json:"id"`
	PaketID        pgtype.UUID    `json:"paket_id"`
	Bulan          int32          `json:"bulan"`
	PersenKeuangan pgtype.Numeric `json:"persen_keuangan"`
	PersenFisik    pgtype.Numeric `json:"persen_fisik"`
}

type RealisasiAnggaranSp2d struct {
	ID          pgtype.UUID        `json:"id"`
	AkunID      pgtype.UUID        `json:"akun_id"`
	Bulan       int32              `json:"bulan"`
	NomorSp2d   string             `json:"nomor_sp2d"`
	TanggalSp2d pgtype.Date        `json:"tanggal_sp2d"`
	NilaiCair   pgtype.Numeric     `json:"nilai_cair"`
	Keterangan  pgtype.Text        `json:"keterangan"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
}

type RevokedToken struct {
	TokenSha256 string             `json:"token_sha256"`
	ExpiresAt   pgtype.Timestamptz `json:"expires_at"`
	RevokedAt   pgtype.Timestamptz `json:"revoked_at"`
}

type User struct {
	ID           pgtype.UUID        `json:"id"`
	Username     string             `json:"username"`
	PasswordHash string             `json:"password_hash"`
	FullName     string             `json:"full_name"`
	Role         string             `json:"role"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
	Permissions  []byte             `json:"permissions"`
}