# backend/openapi.yaml
openapi: 3.1.0
info:
  title: SiAP-PUSBANGKOMSDACKPS API
  version: 1.0.0
  description: API for SiAP-PUSBANGKOMSDACKPS (Sistem Aplikasi Pencairan - PUSBANGKOMSDACKPS)
servers:
  - url: /api/v1
    description: Initial API v1
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    AnggaranNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        jenis:
          type: string
        kode:
          type: string
        uraian:
          type: string
        tahun_anggaran:
          type: integer
        pagu_revisi:
          type: string
        lock_pagu:
          type: string
        realisasi_periode_lalu:
          type: string
        realisasi_periode_ini:
          type: string
        realisasi_sd_periode:
          type: string
        persentase_realisasi:
          type: string
        sisa_anggaran:
          type: string
        level:
          type: integer
        path:
          type: array
          items:
            type: string
    AnggaranImportResult:
      type: object
      properties:
        programs_upserted:
          type: integer
        akun_upserted:
          type: integer
        sp2d_inserted:
          type: integer
    UpdateAnggaranRequest:
      type: object
      properties:
        pagu_revisi:
          type: string
        realisasi_periode_lalu:
          type: string
        realisasi_periode_ini:
          type: string
    UpdateLockPaguRequest:
      type: object
      properties:
        lock_pagu:
          type: string
      required: [lock_pagu]
    SnapshotRequest:
      type: object
      properties:
        periode:
          type: string
          description: "Format YYYY-MM"
      required: [periode]
    AnggaranDokumenItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        anggaran_node_id:
          type: string
          format: uuid
        file_hash_sha256:
          type: string
        original_name:
          type: string
        mime_type:
          type: string
        file_size_bytes:
          type: integer
        created_at:
          type: string
          format: date-time
        uploaded_by:
          type: string
          format: uuid
        uploaded_by_name:
          type: string
    DipaDokumenItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tahun_anggaran:
          type: integer
        bulan:
          type: integer
        revisi:
          type: integer
        file_hash_sha256:
          type: string
        original_name:
          type: string
        mime_type:
          type: string
        file_size_bytes:
          type: integer
        created_at:
          type: string
          format: date-time
        uploaded_by:
          type: string
          format: uuid
        uploaded_by_name:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN_KEUANGAN, PPK, PENGAWAS]
        permissions:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN_KEUANGAN, PPK, PENGAWAS]
        permissions:
          type: array
          items:
            type: string
      required: [username, password, full_name, role, permissions]
    UpdateUserRequest:
      type: object
      properties:
        full_name:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN_KEUANGAN, PPK, PENGAWAS]
        permissions:
          type: array
          items:
            type: string
        password:
          type: string

          enum: [APPROVED, REJECTED]
        rejection_reason:
          type: string
      required: [status]
security:
  - cookieAuth: []
paths:
...
  /anggaran/import:
    post:
      operationId: importAnggaranData
      summary: Import Anggaran data from Excel/CSV
      tags: [Anggaran]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                tahun_anggaran:
                  type: integer
              required:
                - file
                - tahun_anggaran
      responses:
        '200':
          description: Import successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnggaranImportResult'
        '400':
          description: Bad request
  /anggaran/tree:
    get:
      operationId: getAnggaranTree
      summary: Get Anggaran hierarchy tree
      tags: [Anggaran]
      parameters:
        - name: tahun
          in: query
          required: true
          schema:
            type: integer
        - name: periode
          in: query
          required: false
          schema:
            type: string
          description: Snapshot periode string (e.g. 2026-04-Rev). If empty or omitted, returns current live data.
        - name: source
          in: query
          required: false
          schema:
            type: string
          description: Comma-separated list of sources to filter (e.g. rkks,fa_detail,emon). If empty, returns fa_detail.
      responses:
        '200':
          description: Anggaran tree data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnggaranNode'
  /anggaran/snapshots:
    get:
      operationId: getAnggaranSnapshots
      summary: List available snapshot periods for a fiscal year
      tags: [Anggaran]
      parameters:
        - name: tahun
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of snapshot period strings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /anggaran/upload-bukti:
    post:
      operationId: uploadBuktiAnggaran
      summary: Upload a document as proof for an anggaran node
      tags: [Anggaran]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                node_id:
                  type: string
                  format: uuid
              required:
                - file
                - node_id
      responses:
        '200':
          description: Document uploaded successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /anggaran/snapshot:
    post:
      operationId: createAnggaranSnapshot
      summary: Create a snapshot of the current anggaran state
      tags: [Anggaran]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotRequest'
      responses:
        '200':
          description: Snapshot created successfully
        '400':
          description: Bad request
        '500':
          description: Internal Server Error
  /anggaran/{id}/lock:
    put:
      operationId: updateLockPagu
      summary: Update the lock pagu value for a specific node
      tags: [Anggaran]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLockPaguRequest'
      responses:
        '200':
          description: Lock Pagu updated successfully
        '400':
          description: Bad request
        '404':
          description: Node not found
  /anggaran/{id}:
    put:
      operationId: updateAnggaranNode
      summary: Update bottom-level anggaran and rollup
      tags: [Anggaran]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAnggaranRequest'
      responses:
        '200':
          description: Anggaran updated and rolled up successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request
        '404':
          description: Node not found
    delete:
      operationId: deleteAnggaranNode
      summary: Delete an anggaran node and recursively update parents
      tags: [Anggaran]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Anggaran deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request
        '404':
          description: Node not found
  /anggaran/{id}/documents:
    get:
      operationId: getAnggaranDokumenByNode
      summary: Get documents for an anggaran node
      tags: [Anggaran]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnggaranDokumenItem'
  /documents/{id}:
    get:
      operationId: downloadDocument
      summary: Download a document by ID
      tags: [Documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File content
        '404':
          description: Not found
    put:
      operationId: updateDocumentName
      summary: Update the name of a document
      tags: [Documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                original_name:
                  type: string
              required: [original_name]
      responses:
        '200':
          description: Document updated successfully
        '400':
          description: Bad request
    delete:
      operationId: deleteDocument
      summary: Delete a document by ID
      tags: [Documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document deleted successfully
        '404':
          description: Not found
  /anggaran/dipa/upload:
    post:
      operationId: uploadDipaDokumen
      summary: Upload a DIPA/RKKS PDF document
      tags: [Anggaran]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                tahun_anggaran:
                  type: integer
                bulan:
                  type: integer
                revisi:
                  type: integer
              required:
                - file
                - tahun_anggaran
                - bulan
                - revisi
      responses:
        '200':
          description: Document uploaded successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
  /anggaran/dipa/documents:
    get:
      operationId: getDipaDocuments
      summary: List DIPA/RKKS documents with optional filters
      tags: [Anggaran]
      parameters:
        - name: tahun
          in: query
          required: true
          schema:
            type: integer
        - name: bulan
          in: query
          required: false
          schema:
            type: integer
        - name: revisi
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of DIPA documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DipaDokumenItem'
  /anggaran/dipa/documents/{id}:
    delete:
      operationId: deleteDipaDokumen
      summary: Delete a DIPA/RKKS document by ID
      tags: [Anggaran]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document deleted successfully
        '404':
          description: Not found
  /users:
    get:
      operationId: listUsers
      summary: List all users
      tags: [Users]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      operationId: createUser
      summary: Create a new user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    put:
      operationId: updateUser
      summary: Update a user
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      operationId: deleteUser
      summary: Delete a user
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted